Restructure project
Headers are stored in include directory
Scripts are stored in scripts directory
