+++ include/manager.hpp
