+++ source/event.cpp
@@ -1,4 +1,4 @@
-#include "event.hpp"
+#include "../include/event.hpp"
 
 void Event::update()
 {
